2021-09-25 : FEAT/피드백 반영 1. WebMvcConfig의 역할 구분, ViewResolver의 세부역할 빈으로 구분 - 하나의 Config 클래스에서 너무 많은 설정이 있다고 판단하여 작게 구분지었습니다. WebMvcConfiguration 클래스에서 ComponentScan을 통해서 나머지 WebMvc 설정을 들고오게 하려했는데 이러한 방식 피드백 부탁드리겠습니다. - ViewResolverConfiguration 내부의 코드를 빈 등록을 통해서 분할했습니다.
