+++ src/main/java/org/prgms/w3d1/configuration/AppConfiguration.java
@@ -8,6 +8,4 @@ import org.springframework.context.annotation.PropertySource;
 
 @Configuration
 @ComponentScan(basePackages = {"org.prgms.w3d1.model", "org.prgms.w3d1.repository", "org.prgms.w3d1.controller"})
-public class AppConfiguration {
-
-}
+public class AppConfiguration {}
