+++ src/main/java/org/prgms/w3d1/configuration/MessageConverterConfiguration.java
@@ -1,0 +1,19 @@
+package org.prgms.w3d1.configuration;
+
+import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.converter.xml.MarshallingHttpMessageConverter;
+import org.springframework.oxm.xstream.XStreamMarshaller;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+public class MessageConverterConfiguration implements WebMvcConfigurer {
+    @Override
+    public void extendMessageConverters(List<HttpMessageConverter<?>> converters) {
+        var messageConverter = new MarshallingHttpMessageConverter();
+        var xStreamMarshaller = new XStreamMarshaller();
+        messageConverter.setMarshaller(xStreamMarshaller);
+        messageConverter.setUnmarshaller(xStreamMarshaller);
+        converters.add(messageConverter);
+    }
+}
