+++ src/main/java/org/prgms/w3d1/configuration/ResourceHandlerConfiguration.java
@@ -1,0 +1,19 @@
+package org.prgms.w3d1.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.resource.EncodedResourceResolver;
+
+@Configuration
+public class ResourceHandlerConfiguration implements WebMvcConfigurer {
+
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/resources/**")
+                .addResourceLocations("/resources/")
+                .setCachePeriod(60)
+                .resourceChain(true)
+                .addResolver(new EncodedResourceResolver());
+    }
+}
