--- src/main/java/org/prgms/w3d1/configuration/ServletConfiguration.java
@@ -1,70 +1,0 @@
-package org.prgms.w3d1.configuration;
-
-import org.springframework.beans.BeansException;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.converter.HttpMessageConverter;
-import org.springframework.http.converter.xml.MarshallingHttpMessageConverter;
-import org.springframework.oxm.xstream.XStreamMarshaller;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.ViewResolverRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.resource.EncodedResourceResolver;
-import org.thymeleaf.spring5.SpringTemplateEngine;
-import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
-import org.thymeleaf.spring5.view.ThymeleafViewResolver;
-
-import java.util.List;
-
-@EnableWebMvc
-@Configuration
-@EnableTransactionManagement
-@ComponentScan(basePackages = "org.prgms.w3d1")
-public class ServletConfiguration implements WebMvcConfigurer, ApplicationContextAware {
-
-    ApplicationContext applicationContext;
-
-    @Override
-    public void configureViewResolvers(ViewResolverRegistry registry) {
-        var springResourceTemplateResolver = new SpringResourceTemplateResolver();
-        springResourceTemplateResolver.setApplicationContext(applicationContext);
-        springResourceTemplateResolver.setPrefix("/WEB-INF/");
-        springResourceTemplateResolver.setSuffix(".html");
-
-        var springTemplateEngine = new SpringTemplateEngine();
-        springTemplateEngine.setTemplateResolver(springResourceTemplateResolver);
-
-        var thymeleafViewResolver = new ThymeleafViewResolver();
-        thymeleafViewResolver.setTemplateEngine(springTemplateEngine);
-        thymeleafViewResolver.setOrder(1);
-        thymeleafViewResolver.setViewNames(new String[]{"views/*"});
-        registry.viewResolver(thymeleafViewResolver);
-    }
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/resources/**")
-            .addResourceLocations("/resources/")
-            .setCachePeriod(60)
-            .resourceChain(true)
-            .addResolver(new EncodedResourceResolver());
-    }
-
-    @Override
-    public void extendMessageConverters(List<HttpMessageConverter<?>> converters) {
-        var messageConverter = new MarshallingHttpMessageConverter();
-        var xStreamMarshaller = new XStreamMarshaller();
-        messageConverter.setMarshaller(xStreamMarshaller);
-        messageConverter.setUnmarshaller(xStreamMarshaller);
-        converters.add(messageConverter);
-    }
-
-    @Override
-    public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-        this.applicationContext = applicationContext;
-    }
-}
