+++ src/main/java/org/prgms/w3d1/configuration/ViewResolverConfiguration.java
@@ -1,0 +1,56 @@
+package org.prgms.w3d1.configuration;
+
+import org.springframework.beans.BeansException;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewResolverRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.thymeleaf.spring5.SpringTemplateEngine;
+import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
+import org.thymeleaf.spring5.view.ThymeleafViewResolver;
+
+@Configuration
+public class ViewResolverConfiguration implements WebMvcConfigurer, ApplicationContextAware {
+
+    ApplicationContext applicationContext;
+
+    @Bean
+    public SpringResourceTemplateResolver springResourceTemplateResolver() {
+        var springResourceTemplateResolver = new SpringResourceTemplateResolver();
+        springResourceTemplateResolver.setApplicationContext(applicationContext);
+        springResourceTemplateResolver.setPrefix("/WEB-INF/");
+        springResourceTemplateResolver.setSuffix(".html");
+
+        return springResourceTemplateResolver;
+    }
+
+    @Bean
+    public SpringTemplateEngine springTemplateEngine() {
+        var springTemplateEngine = new SpringTemplateEngine();
+        springTemplateEngine.setTemplateResolver(springResourceTemplateResolver());
+
+        return springTemplateEngine;
+    }
+
+    @Bean
+    public ThymeleafViewResolver thymeleafViewResolver() {
+        var thymeleafViewResolver = new ThymeleafViewResolver();
+        thymeleafViewResolver.setTemplateEngine(springTemplateEngine());
+        thymeleafViewResolver.setOrder(1);
+        thymeleafViewResolver.setViewNames(new String[]{"views/*"});
+
+        return thymeleafViewResolver;
+    }
+
+    @Override
+    public void configureViewResolvers(ViewResolverRegistry registry) {
+        registry.viewResolver(thymeleafViewResolver());
+    }
+
+    @Override
+    public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
+        this.applicationContext = applicationContext;
+    }
+}
